client: refactor after review
