+++ client/src/pages/Home/Home.tsx
@@ -1,4 +1,4 @@
-import {ManagmentServiceApiInstance} from 'api/managment-service';
+import {ManagmentServiceApiInstance, ReportDetails} from 'api/managment-service';
 import PageTemplate from 'components/PageTemplate/PageTemplate';
 import HeaderWithIcon from 'components/PageTemplate/components/HeaderWithIcon/HeaderWithIcon';
 import { useEffect, useState } from 'react';
@@ -6,14 +6,15 @@ import ReportDetailsSection from './components/ReportDetailsSection/ReportDetail
 import useReportDetails from 'hooks/useReportStats';
 
 const Home = () => {
-  const [lastReportId, setLastReportId] = useState<string | null>(null);
+  const [lastReport, setLastReport] = useState<ReportDetails | null>(null);
+  const [isReportLoading, setIsReportLoading] = useState(true);
 
   useEffect(() => {
     const getLatestReport = async () => {
       try {
         const latestReport = await ManagmentServiceApiInstance.getLatestReport();
         if (latestReport) {
-          setLastReportId(latestReport.id);
+          setLastReport(latestReport);
         }
       } catch (e: unknown) {
         console.error('Failed to fetch the latest report', e);
@@ -21,20 +22,19 @@ const Home = () => {
     };
 
     getLatestReport();
+    setIsReportLoading(false);
   }, []);
 
   const {
     incidents,
-    report,
     incidentStats,
     areIncidentsLoading,
-    isReportLoading,
-  } = useReportDetails(lastReportId);
+  } = useReportDetails(lastReport);
 
   return (
     <PageTemplate header={<HeaderWithIcon title={'Dashboard'} />}>
       <ReportDetailsSection
-        report={report}
+        report={lastReport}
         incidents={incidents}
         incidentStats={incidentStats}
         areIncidentsLoading={areIncidentsLoading}
