+++ client/src/pages/ReportDetails/ReportDetails.tsx
@@ -1,13 +1,13 @@
 import PageTemplate from 'components/PageTemplate/PageTemplate';
-import {useReportDetails,  IncidentStats } from 'hooks/useReportStats';
-import { useNavigate, useParams } from 'react-router-dom';
+import {useReportDetails, IncidentStats} from 'hooks/useReportStats';
+import {useNavigate, useParams} from 'react-router-dom';
 import './ReportDetails.scss';
 import SectionComponent from 'components/SectionComponent/SectionComponent';
 import SVGIcon from 'components/SVGIcon/SVGIcon';
 import StatisticsDisplay, {
   StatItemData,
 } from 'components/StatisticsDisplay/StatisticsDisplay';
-import { ReportDetails } from 'api/managment-service';
+import {ManagmentServiceApiInstance, ReportDetails} from 'api/managment-service';
 import colors from 'global/colors';
 import IncidentList from 'components/IncidentList/IncidentList';
 import ReportHeader from './components/ReportHeader/ReportHeader';
@@ -16,6 +16,7 @@ import {
   urgencyIncidentCount,
 } from 'types/incident';
 import CenteredSpinner from 'components/CenteredSpinner/CenteredSpinner';
+import {useEffect, useState} from "react";
 
 const statItems = (
   report: ReportDetails,
@@ -102,16 +103,35 @@ const statItems = (
 };
 
 const ReportDetailsPage = () => {
-  const { id } = useParams();
-  const { incidents, report, incidentStats, isReportLoading } =
-    useReportDetails(id!);
+  const {id} = useParams();
+  const [report, setReport] = useState<ReportDetails | null>(null);
+  const [isReportLoading, setIsReportLoading] = useState(true);
+  const {incidents, incidentStats} =
+    useReportDetails(report);
+
+  useEffect(() => {
+    if (!id) {
+      return
+    }
+    const fetchReport = async (id: string) => {
+      try {
+        const reportData = await ManagmentServiceApiInstance.getReport(id);
+        setReport(reportData);
+        setIsReportLoading(false);
+      } catch (e: unknown) {
+        console.error('Failed to fetch report by id', id);
+      }
+    };
+
+    fetchReport(id);
+  }, [id]);
 
   const navigate = useNavigate();
 
   if (isReportLoading || !report || !incidents || !incidentStats) {
     return (
       <PageTemplate header={''}>
-        <CenteredSpinner />
+        <CenteredSpinner/>
       </PageTemplate>
     );
   }
@@ -128,7 +148,7 @@ const ReportDetailsPage = () => {
     >
       <div className="report-details">
         <SectionComponent
-          icon={<SVGIcon iconName={'report-stats-icon'} />}
+          icon={<SVGIcon iconName={'report-stats-icon'}/>}
           title={'Statistics'}
         >
           <StatisticsDisplay
@@ -137,28 +157,28 @@ const ReportDetailsPage = () => {
           />
         </SectionComponent>
         <SectionComponent
-          icon={<SVGIcon iconName={'application-incident-metadata-icon'} />}
+          icon={<SVGIcon iconName={'application-incident-metadata-icon'}/>}
           title={'Application incidents'}
         >
           <IncidentList
             incidents={genericIncidentsFromApplicationIncidents(
               incidents.applicationIncidents,
             )}
-            onClick={({ id: incidentId }) =>
+            onClick={({id: incidentId}) =>
               navigate(`/application-incidents/${incidentId}`)
             }
           />
         </SectionComponent>
 
         <SectionComponent
-          icon={<SVGIcon iconName={'node-incident-metadata-icon'} />}
+          icon={<SVGIcon iconName={'node-incident-metadata-icon'}/>}
           title={'Node incidents'}
         >
           <IncidentList
             incidents={genericIncidentsFromNodeIncidents(
               incidents.nodeIncidents,
             )}
-            onClick={({ id: incidentId }) =>
+            onClick={({id: incidentId}) =>
               navigate(`/node-incidents/${incidentId}`)
             }
           />
